Update SDK targets
Based on the minimum version of Xcode Apple allows to publish to the App Store, the
current minimum iOS SDK target possible is 12 (this is 14's minimum) and as of
April 29, when Apple decommissions Xcode 14, it will be 12.5 (Xcode 15's minimum).

In the interest of code health, best practices and warning reductions, this updates
the SDK targets to use the RECOMMENDED_*_DEPLOYMENT_TARGET build variables introduced
in Xcode 14. They will always provide the minimum SDK that the running Xcode can
build against. This eliminates future maintenance. This PR also eliminates redundant
build settings so that only the project level setting needs to be changed if future
modifications are necessary. Individual targets can still be overridden if they need
a newer minimum target, but not all targets need to be overridden to do so.